On branch mck-keyshortcuts, modified:   aria/aria.html. Removed paragraph: <p>Authors SHOULD avoid using keyboard shortcuts that are used by the user agent. Implementing keyboard shortcuts that conflict with the user agent, will result in a loss of user agent functionality. Authors SHOULD avoid keyboard shortcuts that are used by supported assistive technologies running on the intended operating system platforms. Assistive technologies frequently steal keys at the operating system level before they ever reach a user agents. For the same reason authors SHOULD avoid keyboard shortcuts that are used by the operating system itself.</p>
Removed following sentences from last paragraph:
Authors MUST ensure <code>aria-keyshortcuts</code> are available throughout the document for associated, active elements.
The <code>aria-keyshortcuts</code> is only exposed to the <a>assistive technology</a> when the element it is not hidden from the user or assistive technology by the author. Consequently, authors MUST visibly expose the element so that it may be discovered by a user of an <a>assistive technology</a>.
An example would be a menu item.
If the menu item were to expose a keyboard shortcut it is assumed that as long as the document has focus, pressing that keyboard shortcut will activate that menu item.

Added final paragraph:
<p>
Authors SHOULD avoid implementing shortcut keys that inhibit operating system, user agent, or assistive technology functionality.
This requires the author to carefully consider both which keys to assign and the contexts and conditions in which the keys are available to the user.
For guidance, see the keyboard shortcuts section of the
<cite> <a href="http://www.w3.org/TR/wai-aria-practices-1.1/#kbd_shortcuts"> <abbr title="Accessible Rich Internet Applications">WAI-ARIA</abbr> Authoring Practices Guide</a> </cite> [[WAI-ARIA-PRACTICES]].
</p>